+++ modules/apps/analytics/analytics-settings-web/src/main/java/com/liferay/analytics/settings/web/internal/portal/settings/configuration/admin/display/AnalyticsSyncedSitesConfigurationScreen.java
@@ -15,18 +15,8 @@
 package com.liferay.analytics.settings.web.internal.portal.settings.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
-import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
-import java.io.IOException;
-
-import java.util.Locale;
-
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -36,12 +26,7 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(immediate = true, service = ConfigurationScreen.class)
 public class AnalyticsSyncedSitesConfigurationScreen
-	implements ConfigurationScreen {
-
-	@Override
-	public String getCategoryKey() {
-		return "analytics-cloud";
-	}
+	extends BaseAnalyticsConfigurationScreen {
 
 	@Override
 	public String getKey() {
@@ -49,32 +34,13 @@ public class AnalyticsSyncedSitesConfigurationScreen
 	}
 
 	@Override
-	public String getName(Locale locale) {
-		return LanguageUtil.get(
-			ResourceBundleUtil.getBundle(locale, getClass()), "synced-sites");
+	protected String getJspPath() {
+		return "/edit_synced_sites.jsp";
 	}
 
 	@Override
-	public String getScope() {
-		return "company";
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		try {
-			RequestDispatcher requestDispatcher =
-				_servletContext.getRequestDispatcher("/edit_synced_sites.jsp");
-
-			requestDispatcher.include(httpServletRequest, httpServletResponse);
-		}
-		catch (ServletException se) {
-			throw new IOException(
-				"Unable to render /edit_synced_sites.jsp", se);
-		}
+	protected ServletContext getServletContext() {
+		return _servletContext;
 	}
 
 	@Reference(
